+++ src/checkout/components/CreditCardForm/index.ts
@@ -1,0 +1,1 @@
+export { CreditCardForm as default } from "./CreditCardForm";