+++ src/checkout/components/index.ts
@@ -9,3 +9,4 @@ export { default as UserAddressSelector } from "./UserAddressSelector";
 export {
   default as ShippingUnavailableModal
 } from "./ShippingUnavailableModal";
+export { default as CreditCardForm } from "./CreditCardForm";